Carve task.py into services; kill task_utils
Move postprocess side effects to services/postprocess, viewcount drain to
services/viewcounts, search index rebuild to services/search. Session
runners live in infra/runtime. task.py is a thin re-export shim.
