+++ chafan_core/app/feed.py
@@ -26,7 +26,7 @@ from chafan_core.app.schemas.event import (
     UpvoteQuestionInternal,
     UpvoteSubmissionInternal,
 )
-from chafan_core.app.task_utils import execute_with_broker, execute_with_db
+from chafan_core.app.infra.runtime import execute_with_broker, execute_with_db
 from chafan_core.db.session import SessionLocal
 from chafan_core.utils.base import map_, unwrap
