+++ chafan_core/app/text_analysis.py
@@ -9,7 +9,7 @@ from chafan_core.app.schemas.user import (
     UserEducationExperienceInternal,
     UserWorkExperienceInternal,
 )
-from chafan_core.app.task_utils import execute_with_db
+from chafan_core.app.infra.runtime import execute_with_db
 from chafan_core.db.session import SessionLocal
 from chafan_core.utils.base import dedup, unwrap
