+++ chafan_core/scheduled/deliver_notifications.py
@@ -1,4 +1,4 @@
-from chafan_core.app.task_utils import execute_with_broker
+from chafan_core.app.infra.runtime import execute_with_broker
 from chafan_core.scheduled.lib import deliver_notifications
