+++ chafan_core/scheduled/lib.py
@@ -9,7 +9,7 @@ from sqlalchemy.orm import Session
 from chafan_core.app import crud, models, rep_manager
 from chafan_core.app.infra.request_context import RequestContext
 from chafan_core.app.email_utils import send_notification_email
-from chafan_core.app.task_utils import execute_with_broker, execute_with_db
+from chafan_core.app.infra.runtime import execute_with_broker, execute_with_db
 from chafan_core.db.session import SessionLocal
 from chafan_core.utils.base import EntityType, filter_not_none
